requirements.txt
pandas>=2.0
xlsxwriter>=3.0
pyyaml>=6.0
pytest>=7.0
